backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.7
pandas>=2.1.3
numpy>=1.24.3
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
alembic>=1.12.1